Implement DB structure based on WalletDB_firstversion.png

+++ src/main/java/com/budget/repository/MainCategoryRepository.java
@@ -7,6 +7,5 @@ import com.budget.model.MainCategory;
 // This will be AUTO IMPLEMENTED by Spring into a Bean called mainCategoryRepository
 // CRUD refers Create, Read, Update, Delete
 
-public interface MainCategoryRepository extends CrudRepository<MainCategory, Integer> {
-
+public interface MainCategoryRepository extends CrudRepository<MainCategory, Long> {
 }
